DABOM-359 refactor: webpushsample 패키지를 domain.webpush로 이동
- Feature-Based 패키지 구조 준수 (styleguide 32-57행)
- com.project.webpushsample → com.project.domain.webpush로 이동
- domain 서브패키지를 entity로 변경 (convention 통일)
