+++ src/main/java/com/project/domain/webpush/controller/dto/VapidPublicKey.java
@@ -1,3 +1,3 @@
-package com.project.webpushsample.controller.dto;
+package com.project.domain.webpush.controller.dto;
 
 public record VapidPublicKey(String publicKey) {}
