+++ src/main/java/com/project/domain/webpush/repository/SubscriptionRepository.java
@@ -1,10 +1,10 @@
-package com.project.webpushsample.repository;
+package com.project.domain.webpush.repository;
 
 import java.util.Optional;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-import com.project.webpushsample.domain.Subscription;
+import com.project.domain.webpush.entity.Subscription;
 
 public interface SubscriptionRepository extends JpaRepository<Subscription, Long> {
